Migrate to using v2 facilities_api in VAOS
