+++ src/applications/vaos/services/mocks/var/cc_providers.json
@@ -710,10 +710,10 @@
     }
   },
   "links": {
-    "self": "https://platform-api.va.gov/facilities_api/v1/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=1&per_page=15&specialties%5B%5D=363LP2300X",
-    "first": "https://platform-api.va.gov/facilities_api/v1/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=1&per_page=15&specialties%5B%5D=363LP2300X",
+    "self": "https://platform-api.va.gov/facilities_api/v2/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=1&per_page=15&specialties%5B%5D=363LP2300X",
+    "first": "https://platform-api.va.gov/facilities_api/v2/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=1&per_page=15&specialties%5B%5D=363LP2300X",
     "prev": null,
-    "next": "https://platform-api.va.gov/facilities_api/v1/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=2&per_page=15&specialties%5B%5D=363LP2300X",
-    "last": "https://platform-api.va.gov/facilities_api/v1/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=2&per_page=15&specialties%5B%5D=363LP2300X"
+    "next": "https://platform-api.va.gov/facilities_api/v2/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=2&per_page=15&specialties%5B%5D=363LP2300X",
+    "last": "https://platform-api.va.gov/facilities_api/v2/ccp/provider?latitude=38.9072&longitude=77.0369&radius=60&bbox%5B%5D=-77.7866&bbox%5B%5D=38.145&bbox%5B%5D=-76.2866&bbox%5B%5D=39.645&page=2&per_page=15&specialties%5B%5D=363LP2300X"
   }
 }
