+++ src/applications/vaos/services/vaos/index.js
@@ -127,3 +127,20 @@ export function getCommunityCareV2(typeOfCare) {
     `/vaos/v2/community_care/eligibility/${typeOfCare}`,
   ).then(parseApiObject);
 }
+
+export async function getCommunityCareFacilities({
+  latitude,
+  longitude,
+  radius,
+  bbox,
+  specialties,
+  page = 1,
+  perPage = 10,
+}) {
+  const bboxQuery = bbox.map(c => `bbox[]=${c}`).join('&');
+  const specialtiesQuery = specialties.map(s => `specialties[]=${s}`).join('&');
+
+  return apiRequestWithUrl(
+    `/facilities_api/v2/ccp/provider?latitude=${latitude}&longitude=${longitude}&radius=${radius}&per_page=${perPage}&page=${page}&${bboxQuery}&${specialtiesQuery}&trim=true`,
+  ).then(parseApiList);
+}
