+++ src/applications/vaos/tests/e2e/page-objects/CommunityCarePreferencesPageObject.js
@@ -51,7 +51,7 @@ export class CommunityCarePreferencesPageObject extends PageObject {
     cy.findByText(label, { selector: 'h2' }).click({
       waitForAnimations: true,
     });
-    cy.wait('@v1:get:provider');
+    cy.wait('@v2:get:provider');
     cy.findByLabelText(/doe, jane/i).check({ waitForAnimations: true });
     cy.findByText(/Select provider/i, { selector: 'button' }).click({
       waitForAnimations: true,
@@ -67,7 +67,7 @@ export class CommunityCarePreferencesPageObject extends PageObject {
     cy.findByText(label, { selector: 'h2' }).click({
       waitForAnimations: true,
     });
-    cy.wait('@v1:get:provider');
+    cy.wait('@v2:get:provider');
     cy.findByLabelText(/doe, jane/i).check({ waitForAnimations: true });
     cy.findByText(/Select provider/i, { selector: 'button' }).click({
       waitForAnimations: true,
